petite redirection sur les fichier php que l'utilisateur ne doit pas acceder

+++ deleteAccount.php
@@ -1,16 +1,21 @@
 <?php
 require_once("includes/autoload.inc.php");
 require_once("includes/myPDO.inc.php");
-require_once("includes/userRestricted.inc.php");
+require_once('includes/connectedMember.inc.php');
+require_once('includes/utility.inc.php');
 
-$member = Member::getInstance();
+if(verify($_SERVER,'HTTP_REFERER')) {
+    $member = Member::getInstance();
 
-$pdo = MyPDO::GetInstance();
-$stmt = $pdo->prepare(<<<SQL
+    $pdo = MyPDO::GetInstance();
+    $stmt = $pdo->prepare(<<<SQL
 			DELETE FROM Membre
             WHERE idMembre=:id;
 SQL
-);
-$stmt->execute(array("id"=>$member->getId()));
-Member::disconnect();
-header('Location: index.php');
+    );
+    $stmt->execute(array("id" => $member->getId()));
+    Member::disconnect();
+    header('Location: index.php');
+}
+else
+    header('Location: index.php');
